pyproject.toml
[project]
name = "macondo-be"
version = "0.1.0"
description = ""
authors = [
    {name = "ivanbahrii"}
]
readme = "README.md"
requires-python = ">=3.10"

[tool.poetry.dependencies]
python = ">=3.10,<4.0"

#api
fastapi = "^0.111.0"
uvicorn = { extras = ["standard"], version = "^0.30.1" }

#ai
langchain = ">=0.3,<0.4"
langchain-core = ">=0.3,<0.4"
langgraph = ">=0.2.20,<0.3"

#util
pydantic = ">=2.0,<3.0"
python-dotenv = "^1.0.1"
chromadb = "^1.0.16"
sentence-transformers = "^5.1.0"
pypdf = "^6.0.0"
langchain-experimental = "^0.3.4"
langchain-ollama = "^0.3.7"
finnhub-python = "^2.4.24"
newspaper3k = "^0.2.8"
lxml-html-clean = "^0.4.2"
bs4 = "^0.0.2"
wikipedia-api = "^0.8.1"
trulens = "^2.4.0"
trulens-apps-langchain = "^2.4.0"
trulens-providers-litellm = "^2.4.0"
litellm = "^1.77.3"
yfinance = "^0.2.66"
boto3 = "^1.40.48"
langchain-aws = "^0.2.35"
langchain-mcp-adapters = "^0.1.11"
docling = "^2.59.0"
loguru = "^0.7.3"
evaluate = "0.4.0"
pyarrow = "20.0.0"
bert-score = "^0.3.13"
guardrails-ai = "^0.6.7"
pytest = "^8.4.2"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
